# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Enterprise Deal Commander (EDC) Phase 1 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v2/activity:
    get:
      operationId: listPortfolioActivity
      tags: [history]
      parameters:
        - { name: since, in: query, required: false, schema: { type: string } }
        - { name: until, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer } }
        - { name: offset, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: Portfolio-wide activity stream (all active deals, newest first)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"
  /v2/deals/{dealId}/activity:
    get:
      operationId: listDealActivity
      tags: [history]
      parameters:
        - { name: dealId, in: path, required: true, schema: { type: string } }
        - { name: event_type, in: query, required: false, schema: { type: string } }
        - { name: since, in: query, required: false, schema: { type: string } }
        - { name: until, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer } }
        - { name: offset, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: Deal activity stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"
...
  /v2/deals/{dealId}/snapshots:
    get:
      operationId: listDealSnapshots
      tags: [history]
      parameters:
        - { name: dealId, in: path, required: true, schema: { type: string } }
        - { name: since, in: query, required: false, schema: { type: string } }
        - { name: until, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer } }
        - { name: offset, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: Deal snapshot list (metadata only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotListResponse"
  /v2/snapshots/{snapshotId}:
    get:
      operationId: getDealSnapshot
      tags: [history]
      parameters:
        - { name: snapshotId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Full snapshot with payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotDetailResponse"

  # ============ GOVERNANCE ============
...
components:
  schemas:
...
    ListMeta:
      type: object
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
      required: [total, limit, offset]
...
    ActivityEvent:
      type: object
      properties:
        id: { type: string }
        dealId: { type: string }
        dealName: { type: ["string", "null"] }
        eventType: { type: string }
        entityType: { type: string }
        entityId: { type: ["string", "null"] }
        summary: { type: string }
        metadata: { type: ["object", "null"], additionalProperties: true }
        actor: { type: string }
        occurredAt: { type: string }
      required:
        [id, dealId, eventType, entityType, summary, actor, occurredAt]
    ActivityListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/ActivityEvent" }
        meta: { $ref: "#/components/schemas/ListMeta" }
      required: [data, meta]
...
    SnapshotSummary:
      type: object
      properties:
        id: { type: string }
        dealId: { type: string }
        reason: { type: string }
        triggerEvent: { type: ["string", "null"] }
        healthStatus: { type: string }
        salesStageId: { type: ["integer", "null"] }
        salesStage: { type: ["string", "null"] }
        calculatedTcv: { type: ["string", "null"] }
        normalizedTcv: { type: ["string", "null"] }
        createdBy: { type: string }
        snapshotAt: { type: string }
      required: [id, dealId, reason, healthStatus, createdBy, snapshotAt]
    SnapshotListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/SnapshotSummary" }
        meta: { $ref: "#/components/schemas/ListMeta" }
      required: [data, meta]
    SnapshotDetail:
      allOf:
        - $ref: "#/components/schemas/SnapshotSummary"
        - type: object
          properties:
            payload: { type: object, additionalProperties: true }
          required: [payload]
    SnapshotDetailResponse:
      type: object
      properties:
        data: { $ref: "#/components/schemas/SnapshotDetail" }
      required: [data]
...